refactor: add /model folder for server and client models
