+++ packages/examples/src/common/language-client-runner.ts
@@ -2,8 +2,9 @@
  * Copyright (c) 2018-2022 TypeFox GmbH (http://www.typefox.io). All rights reserved.
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
+
+import { LanguageClientRunConfig } from '../model/client.js';
 import { createMonacoEditor, createUrl, doInit, initWebSocketAndStartClient } from './client-commons.js';
-import { LanguageClientRunConfig } from './model.js';
 /** LSP client runner */
 export const runLanguageClient = async (config : LanguageClientRunConfig) => {
     const languageId = config.registerConfig.id;
