+++ packages/examples/src/groovy/server/main.ts
@@ -6,14 +6,14 @@
 import { runLanguageServer } from '../../common/language-server-runner.js';
 import { resolve } from 'path';
 import { groovyConfig } from '../config.js';
-import { LanguageCli } from '../../common/model.js';
+import { LanguageName } from '../../model/shared.js';
 export const runGroovyLanguageServer = (baseDir: string, relativeDir: string) => {
     const processRunPath = resolve(baseDir, relativeDir);
     runLanguageServer({
         serverName: 'GROOVY',
         pathName: groovyConfig.path,
         serverPort: groovyConfig.port,
-        runCommand: LanguageCli.java,
+        runCommand: LanguageName.java,
         runCommandArgs: [
             '-jar',
             processRunPath
