+++ packages/examples/src/json/server/main.ts
@@ -4,7 +4,7 @@
  * ------------------------------------------------------------------------------------------ */
 import { runLanguageServer } from '../../common/language-server-runner.js';
 import { resolve } from 'path';
-import { LanguageCli } from '../../common/model.js';
+import { LanguageName } from '../../model/shared.js';
 
 export const runJsonServer = (baseDir: string, relativeDir: string) => {
     const processRunPath = resolve(baseDir, relativeDir);
@@ -12,7 +12,7 @@ export const runJsonServer = (baseDir: string, relativeDir: string) => {
         serverName: 'JSON',
         pathName: '/sampleServer',
         serverPort: 30000,
-        runCommand: LanguageCli.node,
+        runCommand: LanguageName.node,
         runCommandArgs: [
             processRunPath,
             '--stdio'
