+++ packages/examples/src/python/server/main.ts
@@ -6,7 +6,7 @@
 import { resolve } from 'path';
 import { runLanguageServer } from '../../common/language-server-runner.js';
 import { IncomingMessage } from 'http';
-import { LanguageCli } from '../../common/model.js';
+import { LanguageName } from '../../model/shared.js';
 
 export const runPythonServer = (baseDir: string, relativeDir: string) => {
     const processRunPath = resolve(baseDir, relativeDir);
@@ -14,7 +14,7 @@ export const runPythonServer = (baseDir: string, relativeDir: string) => {
         serverName: 'PYRIGHT',
         pathName: '/pyright',
         serverPort: 30001,
-        runCommand: LanguageCli.node,
+        runCommand: LanguageName.node,
         runCommandArgs: [
             processRunPath,
             '--stdio'
